feat: 레이스 목표를 소수 km로 저장 (마일 무손실)
- V17: challenge.goal_km integer → numeric(10,3) (미터급 정밀도)
- Challenge.goalKm BigDecimal화 + 검증(signum/compareTo)·진척도·DTO 4종 정리
- 목표 입력 소수점 허용, 현재 단위로 입력받아 km로 환산 저장(setScale 3)
- 수정 화면은 저장된 km를 현재 단위로 환산해 프리필, 클램프·안내도 단위 기준
- 목표 문구에서 km/정수 제거(5개 언어), 라벨에 단위 동적 표시

+++ backend/src/main/java/com/runrace/backend/challenge/dto/ChallengeDetailResponse.java
@@ -1,12 +1,13 @@
 package com.runrace.backend.challenge.dto;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
 
 public record ChallengeDetailResponse(
     Long id,
     String title,
-    int goalKm,
+    BigDecimal goalKm,
     int maxMembers,
     String startAt,
     String endAt,
